Extract a base class with most of the implementation.
